admin/requirements.txt
asgiref==3.5.0
click==8.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
pika==1.2.0
pytz==2021.3
sqlparse==0.4.2
Werkzeug==2.0.3
